input: misc: aw8697_haptic: Cast 'effect->pattern' to smaller integer type

+++ drivers/input/misc/aw8697_haptic/aw8697.c
@@ -4358,7 +4358,7 @@ static int aw8697_parse_dt_common(struct device *dev, struct aw8697 *aw8697,
 		}
 		printk("%s %d effect->pattern_length=%d effect->pattern=%p\n",
 		       __func__, __LINE__, effect->pattern_length,
-		       effect->pattern);
+		       (int)(unsigned long)effect->pattern);
 
 		effect->play_rate_us = config->play_rate_us;
 		rc = of_property_read_u32(child_node, "qcom,wf-play-rate-us",
